+++ internal/cmd/marketplace/list_version_test.go
@@ -16,10 +16,13 @@
 package marketplace
 
 import (
+	"encoding/json"
+	"fmt"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
+	"github.com/mia-platform/miactl/internal/client"
 	"github.com/mia-platform/miactl/internal/clioptions"
 	"github.com/mia-platform/miactl/internal/resources/marketplace"
 	"github.com/stretchr/testify/assert"
@@ -38,7 +41,18 @@ const (
 		"visibility": {
 		  "public": true
 		}
-	  }
+	},
+	{
+		"name": "Some Awesome Service v2",
+		"description": "The Awesome Service allows to do some amazing stuff.",
+		"version": "2.0.0",
+		"reference": "655342ce0f991db238fd73e4",
+		"security": false,
+		"releaseNote": "-",
+		"visibility": {
+		  "public": true
+		}
+	}
 ]`
 )
 
@@ -48,6 +62,100 @@ func TestNewListVersionsCmd(t *testing.T) {
 		cmd := ListVersionCmd(opts)
 		require.NotNil(t, cmd)
 	})
+}
+
+func TestGetItemVersions(t *testing.T) {
+	testCases := []struct {
+		testName string
+
+		companyID string
+		itemID    string
+
+		statusCode    int
+		errorResponse map[string]string
+
+		expected    []marketplace.Release
+		expectedErr error
+	}{
+		{
+			testName:   "should return correct result when the endpoint answers 200 OK",
+			companyID:  "some-company",
+			itemID:     "some-item",
+			statusCode: http.StatusOK,
+			expected: []marketplace.Release{
+				{
+					Name:        "Some Awesome Service",
+					Description: "The Awesome Service allows to do some amazing stuff.",
+					Version:     "1.0.0",
+				},
+				{
+					Name:        "Some Awesome Service v2",
+					Description: "The Awesome Service allows to do some amazing stuff.",
+					Version:     "2.0.0",
+				},
+			},
+			expectedErr: nil,
+		},
+		{
+			testName:   "should return not found error if item is not found",
+			companyID:  "some-company",
+			itemID:     "some-item",
+			statusCode: http.StatusNotFound,
+			errorResponse: map[string]string{
+				"error": "Not Found",
+			},
+			expected:    nil,
+			expectedErr: ErrItemVersionNotFound,
+		},
+		{
+			testName:   "should return generic error if item is not found",
+			companyID:  "some-company",
+			itemID:     "some-item",
+			statusCode: http.StatusInternalServerError,
+			errorResponse: map[string]string{
+				"error": "Internal Server Error",
+			},
+			expected:    nil,
+			expectedErr: ErrGenericItemVersion,
+		},
+	}
+
+	for _, testCase := range testCases {
+		t.Run(testCase.testName, func(t *testing.T) {
+			var expectedBytes []byte
+			if testCase.expected != nil {
+				expectedBytes = []byte(listVersionsMockResponseBody)
+			} else {
+				var err error
+				expectedBytes, err = json.Marshal(testCase.errorResponse)
+				require.NoError(t, err)
+			}
+
+			mockServer := buildMockListVersionServer(
+				t,
+				testCase.statusCode,
+				expectedBytes,
+				testCase.companyID,
+				testCase.itemID,
+			)
+			defer mockServer.Close()
+			client, err := client.APIClientForConfig(&client.Config{
+				Transport: http.DefaultTransport,
+				Host:      mockServer.URL,
+			})
+			require.NoError(t, err)
+
+			found, err := getItemVersions(client, testCase.companyID, testCase.itemID)
+			if testCase.expectedErr != nil {
+				require.ErrorIs(t, err, testCase.expectedErr)
+				require.Nil(t, found)
+			} else {
+				require.NoError(t, err)
+				require.Equal(t, &testCase.expected, found)
+			}
+		})
+
+	}
 }
 
 func TestBuildMarketplaceItemVersionList(t *testing.T) {
@@ -72,7 +180,7 @@ func TestBuildMarketplaceItemVersionList(t *testing.T) {
 
 	for testName, testCase := range testCases {
 		t.Run(testName, func(t *testing.T) {
-			found, err := buildItemVersionList(testCase.releases)
+			found, err := buildItemVersionList(&testCase.releases)
 			assert.NoError(t, err)
 			assert.NotZero(t, found)
 			for _, expected := range testCase.expectedContains {
@@ -82,20 +190,17 @@ func TestBuildMarketplaceItemVersionList(t *testing.T) {
 	}
 }
 
-func mockListVersionsServer(t *testing.T, validResponse bool) *httptest.Server {
+func buildMockListVersionServer(t *testing.T, statusCode int, responseBody []byte, companyID, itemID string) *httptest.Server {
 	t.Helper()
 
 	return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.RequestURI != listMarketplaceEndpoint && r.Method != http.MethodGet {
-			w.WriteHeader(http.StatusNotFound)
-			require.Fail(t, "unsupported call")
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-		if validResponse {
-			w.Write([]byte(mockResponseBody))
-			return
-		}
-		w.Write([]byte(`{"message": "invalid json"}`))
+		require.Equal(t, http.MethodGet, r.Method)
+		require.Equal(
+			t,
+			r.RequestURI,
+			fmt.Sprintf(listItemVersionsEndpointTemplate, companyID, itemID),
+		)
+		w.WriteHeader(statusCode)
+		w.Write(responseBody)
 	}))
 }